[FEAT]: 투표 마감 옵션 및 날짜 선택 검증 추가
